Adding manual bot

+++ src/main/java/com/pm/mentor/darkforest/ai/GameActionApi.java
@@ -6,7 +6,7 @@ import com.loxon.javachallenge.challenge.game.event.action.GameAction;
 public interface GameActionApi {
 	GameAction spaceMission(int sourcePlanet, int targetPlanet);
 	GameAction spaceMissionWithWormHole(int sourcePlanet, int targetPlanet, int wormHole, EntryPointIndex wormHoleSide);
-	GameAction buildWormHole(int xa, int ya, int xb, int yb);
+	GameAction buildWormHole(long xa, long ya, long xb, long yb);
 	GameAction erectShield(int targetPlanet);
 	GameAction shootMBH(int sourcePlanet, int targetPlanet);
 }
